Move DEstinationHeaderSectionViewModel to its own file and make it implements Observer pattern
